Add "About" page with library information table and update menu navigation

+++ covadis-webserver/src/main/groovy/org/noise_planet/covadis/scripts/Slurm/Noise_level_from_source.groovy
@@ -323,7 +323,7 @@ def exec(DataSource dataSource, Map inputs, ProgressVisitor progress) {
                 sshKeyArmoredString: res.private_key,
                 sshKeyPassword: inputs.key_password,
                 user: res.user_name,
-                maxTasksPerJobs: inputs.getOrDefault("slurm_task_count", 8),
+                maxTasksPerJobs: inputs.getOrDefault("slurm_task_count", 8) as Integer,
                 serverKey: res.ssl_key,
                 serverKeyType: res.ssh_key_type,
                 javaBinaryPath: res.java_binary_path)
